refactor: introduce @bitwarden/platform-utils
