+++ .github/CODEOWNERS
@@ -104,6 +104,7 @@ libs/state @bitwarden/team-platform-dev
 libs/state-test-utils @bitwarden/team-platform-dev
 libs/device-type @bitwarden/team-platform-dev
 libs/encoding @bitwarden/team-platform-dev
+libs/platform-utils @bitwarden/team-platform-dev
 # Web utils used across app and connectors
 apps/web/src/utils/ @bitwarden/team-platform-dev
 # Web core and shared files
