+++ libs/platform-utils/README.md
@@ -1,0 +1,5 @@
+# platform-utils
+
+Owned by: platform
+
+Assorted utils for working with toasts, device type, clipboard, etc. This library needs to be broken up
