+++ libs/platform-utils/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@bitwarden/platform-utils",
+  "version": "0.0.1",
+  "description": "Assorted utils for working with toasts, device type, clipboard, etc. This library needs to be broken up",
+  "private": true,
+  "type": "commonjs",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "license": "GPL-3.0",
+  "author": "platform"
+}
