+++ libs/platform-utils/src/platform-utils.spec.ts
@@ -1,0 +1,8 @@
+import * as lib from "./index";
+
+describe("platform-utils", () => {
+  // This test will fail until something is exported from index.ts
+  it("should work", () => {
+    expect(lib).toBeDefined();
+  });
+});
